# perf: optimize channel buffer and object pooling