- Change listener's pool type from proto.Message to []proto.Field as we only use the pool for Fields, this will reduce the memory used by the pool channel since sizeof `proto.Message` is 56 bytes where a slice is only 24 bytes.
- Pick optimal value for default channel buffer based on benchmark, we have few candidates [8, 16, 24, 32, 64, 128, 256, 512, 1000 (previous default)], the values are actually just arbitrary number, it does not have special meaning of why it's a multiple of 8. Let's just take a look on these benchmark result:
```js
goos: darwin | goarch: amd64 | pkg: benchfit
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
                       │     8.txt     │                16.txt                │               32.txt                │                64.txt                │               128.txt               │               256.txt               │               512.txt               │              1000.txt               │
                       │    sec/op     │    sec/op     vs base                │   sec/op     vs base                │    sec/op     vs base                │   sec/op     vs base                │   sec/op     vs base                │   sec/op     vs base                │   sec/op     vs base                │
Decode/muktihari/fit-4   186.01m ± 29%   134.35m ± 3%  -27.77% (p=0.000 n=10)   92.21m ± 2%  -50.43% (p=0.000 n=10)   92.52m ± 12%  -50.26% (p=0.000 n=10)   90.08m ± 1%  -51.57% (p=0.000 n=10)   90.54m ± 2%  -51.33% (p=0.000 n=10)   89.50m ± 3%  -51.89% (p=0.000 n=10)   91.23m ± 7%  -50.95% (p=0.000 n=10)

                       │    8.txt     │               16.txt                │               32.txt                │               64.txt                │               128.txt               │               256.txt               │               512.txt               │              1000.txt               │
                       │     B/op     │     B/op      vs base               │     B/op      vs base               │     B/op      vs base               │     B/op      vs base               │     B/op      vs base               │     B/op      vs base               │     B/op      vs base               │
Decode/muktihari/fit-4   49.73Mi ± 0%   49.73Mi ± 0%  +0.01% (p=0.000 n=10)   49.74Mi ± 0%  +0.02% (p=0.000 n=10)   49.75Mi ± 0%  +0.06% (p=0.000 n=10)   49.79Mi ± 0%  +0.12% (p=0.000 n=10)   49.85Mi ± 0%  +0.25% (p=0.000 n=10)   49.98Mi ± 0%  +0.51% (p=0.000 n=10)   50.23Mi ± 0%  +1.01% (p=0.000 n=10)

                       │    8.txt    │               16.txt               │               32.txt               │               64.txt               │              128.txt               │              256.txt               │              512.txt               │              1000.txt              │
                       │  allocs/op  │  allocs/op   vs base               │  allocs/op   vs base               │  allocs/op   vs base               │  allocs/op   vs base               │  allocs/op   vs base               │  allocs/op   vs base               │  allocs/op   vs base               │
Decode/muktihari/fit-4   100.1k ± 0%   100.1k ± 0%  +0.01% (p=0.000 n=10)   100.1k ± 0%  +0.02% (p=0.000 n=10)   100.1k ± 0%  +0.06% (p=0.000 n=10)   100.2k ± 0%  +0.12% (p=0.000 n=10)   100.3k ± 0%  +0.25% (p=0.000 n=10)   100.6k ± 0%  +0.50% (p=0.000 n=10)   101.1k ± 0%  +0.99% (p=0.000 n=10)
```

The listener performance with channel buffers sizes 8 and 16 suffers, while starting from 32, the performance seems to be significantly faster: `32 and 64 are around 92ms`, while `128, 256, 512 and 1000 are around 90ms`. It seems having buffer more than 128 does not add more value, rather it increases memory allocation. From the benchmark result, `we pick 128 as default value`, it should be sufficient for most cases. Users can still change the value via option as needed.

Note:  The way we do benchmark have limitations, the machine may take big influence on the resulting number as well as the FIT file used for benchmarking and the Go version used (v1.20.8). We are open to suggestions regarding this matter, as people may find that there is a more optimal value.